refactored spawner into 2 seperate spawner for neutral and police, change spawn method to zone-based
